pyproject.toml
[tool.poetry]
name = "b2b-content-agent-system"
version = "1.0.0"
description = "Multi-Agent B2B Content Generation System with Human-in-the-Loop (HITL) Orchestration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/yourusername/content-agent-system"
repository = "https://github.com/yourusername/content-agent-system"
documentation = "https://github.com/yourusername/content-agent-system/blob/main/HITL_GUIDE.md"
keywords = [
    "ai",
    "crewai",
    "content-generation",
    "b2b-marketing",
    "hitl",
    "gemini",
    "multi-agent-system"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Marketing",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Marketing",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
packages = [
    { include = "b2b_content_agent", from = "src" }
]

[tool.poetry.scripts]
# HITL Pipeline
hitl-run = "b2b_content_agent.run_hitl:main"
# Main entry point (for backward compatibility)
content-agent = "src.main:main"

[tool.poetry.dependencies]
python = "^3.11"
# Core AI Framework
crewai = { version = "^1.3.0", extras = ["google-genai"] }
crewai-tools = "0.11.0"
google-generativeai = "^0.8.0"
# Memory & State Management
mem0ai = "^0.1.0"
# API & Web Framework (optional, for future web UI)
fastapi = { version = "^0.115.0", optional = true }
uvicorn = { version = "^0.32.0", optional = true }
# Configuration & Environment
python-dotenv = "^1.0.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
# Document Processing
pypdf2 = "^3.0.0"
python-docx = "^1.1.0"
# Web Scraping & APIs
beautifulsoup4 = "^4.12.0"
playwright = "^1.48.0"
requests = "^2.32.0"
# LangChain Integration
langchain = "^0.3.0"
langchain-google-genai = "^2.0.0"
# CLI Tools
typer = { version = "^0.12.0", optional = true }
rich = { version = "^13.0.0", optional = true }

[tool.poetry.extras]
# Optional feature groups
api = ["fastapi", "uvicorn"]
cli = ["typer", "rich"]
all = ["fastapi", "uvicorn", "typer", "rich"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
isort = "^5.13.0"
mypy = "^1.13.0"
ruff = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--disable-warnings",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.black]
line-length = 100
target-version = ['py311', 'py312', 'py313']
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["b2b_content_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
